End to end test to read from SQL and write to LevelDB.
Render rows from SQL as []map[string]interface{}

For each row we need to allocate a scannable []interface{} and the
map[string]interface{} that we return.

The map has a special key "__pk" to store a string representation of the
row's primary key.

LevelDB is a test global, and it should write it's state under /tmp.